Rename `app-framework` -> `sys-framework`.
