+++ src/builtin-applications/export/StaticFiles.js
@@ -4,7 +4,7 @@
 import fs from 'node:fs/promises';
 
 import { ApplicationConfig } from '@this/sys-config';
-import { BaseApplication } from '@this/app-framework';
+import { BaseApplication } from '@this/sys-framework';
 import { Paths, Statter } from '@this/fs-util';
 import { IntfLogger } from '@this/loggy';
 import { EtagGenerator, HttpUtil, MimeTypes } from '@this/net-util';
