+++ src/builtin-services/export/MemoryMonitor.js
@@ -5,7 +5,7 @@ import { memoryUsage } from 'node:process';
 import { setTimeout } from 'node:timers/promises';
 
 import { ServiceConfig } from '@this/sys-config';
-import { BaseService } from '@this/app-framework';
+import { BaseService } from '@this/sys-framework';
 import { Threadlet } from '@this/async';
 import { Moment } from '@this/data-values';
 import { Host } from '@this/host';
