+++ src/builtin-services/export/ProcessInfoFile.js
@@ -5,7 +5,7 @@ import * as fs from 'node:fs/promises';
 import * as timers from 'node:timers/promises';
 
 import { FileServiceConfig } from '@this/sys-config';
-import { BaseService } from '@this/app-framework';
+import { BaseService } from '@this/sys-framework';
 import { Saver } from '@this/app-util';
 import { Threadlet } from '@this/async';
 import { Duration, Moment } from '@this/data-values';
