+++ src/builtin-services/package.json
@@ -13,7 +13,7 @@
 
   "dependencies": {
     "@this/sys-config": "*",
-    "@this/app-framework": "*",
+    "@this/sys-framework": "*",
     "@this/app-util": "*",
     "@this/async": "*",
     "@this/data-values": "*",
