+++ src/sys-framework/export/ComponentManager.js
