+++ src/sys-framework/export/ComponentRegistry.js
