+++ src/sys-framework/export/NetworkEndpoint.js
