+++ src/sys-framework/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/app-framework",
+  "name": "@this/sys-framework",
   "version": "0.0.1",
   "type": "module",
   "private": true,
