+++ src/sys-framework/private/ThisModule.js
